+++ packages/editor-preview/src/browser/editor-preview-frontend-module.ts
@@ -17,7 +17,7 @@
 import '../../src/browser/style/index.css';
 import { FrontendApplicationContribution, KeybindingContribution, WidgetFactory } from '@theia/core/lib/browser';
 import { ContainerModule } from '@theia/core/shared/inversify';
-import { bindEditorPreviewPreferences } from './editor-preview-preferences';
+import { bindEditorPreviewPreferences } from '../common/editor-preview-preferences';
 import { EditorPreviewManager } from './editor-preview-manager';
 import { EditorManager } from '@theia/editor/lib/browser';
 import { EditorPreviewWidgetFactory } from './editor-preview-widget-factory';